Another solution for the DnD wxWidgets issue  (also affecting Wrye Bash) (#84)
caused by GMEM_MOVEABLE GlobalAlloc memory. This patch operate on
RtlSizeHeap allowing it to return properly the allocated size of this kind of
memory.  This is basically adding a retry on the first error path, with
the pointer -sizeof(HGLOBAL_STORAGE)*2. Then this size is also
subtracted to the return value